factor taskdb into multiple modules
